# src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "v1",
    "title": "Store Front HQ",
    "description": ""
  },
  "paths": {
...
    "/admin/user/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get a user.",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "Update a user.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "in": "path",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Remove a user.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "description": "Fields to update a user",
            "required": [
              "id"
            ]
          }
        ]
      },
      "User": {
        "allOf": [
          {
            "type": "object",
            "description": "A Store Front HQ user",
            "properties": {
              "id": {
                "type": "string",
                "description": "the user id"
              }
            }
          },
          {
            "$ref": "#/components/schemas/BaseUser"
          }
        ]
      },
      "BaseUser": {
        "type": "object",
        "description": "Base user fields",
        "properties": {
          "username": {
            "type": "string",
            "description": "user name"
          },
          "email": {
            "type": "string",
            "description": "user email"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "httpCode": {
            "type": "integer",
            "description": "The HTTP status code."
          },
          "eventId": {
            "type": "string",
            "description": "An id to identify the request and response transaction."
          },
          "errorCode": {
            "type": "integer",
            "description": "The service-specific error code."
          },
          "message": {
            "type": "string",
            "description": "The reason for the error."
          },
          "detail": {
            "type": "string",
            "description": "Optional details on the cause of this error."
          }
        }
      }
    }
  }
}